refactor: Email 도메인 패키지 변경
