+++ src/main/java/com/example/demo/domain/email/EmailSender.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain;
+package com.example.demo.domain.email;
 
 public interface EmailSender {
