+++ src/main/java/com/example/demo/service/MailServiceImpl.java
@@ -4,9 +4,9 @@ import static com.example.demo.exception.ExceptionType.*;
 
 import org.springframework.stereotype.Service;
 
-import com.example.demo.domain.EmailSender;
-import com.example.demo.domain.EmailTemplate;
 import com.example.demo.domain.MemberImpl;
+import com.example.demo.domain.email.EmailSender;
+import com.example.demo.domain.email.EmailTemplate;
 import com.example.demo.dto.EmailSendRequest;
 import com.example.demo.dto.EmailTemplateResponse;
 import com.example.demo.exception.CafegoryException;
